requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic-settings==2.7.1
openai==1.59.3
scikit-learn==1.6.1
pandas==2.2.3
joblib==1.4.2
jinja2==3.1.5
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
